UPDATE mmnotice saver

See merge request s07-webmobile2-sub2/S07P12E204!95

Former-commit-id: c6eb5c51ff9e9a1b39fb6193efeaeb067201d0e3

+++ Back-End/ssachedule/src/main/java/com/ssljjong/ssachedule/repository/LunchRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface LunchRepository extends JpaRepository<Lunch, Long> {
+public interface LunchRepository extends JpaRepository<Lunch, Integer> {
 
 
     List<Lunch> findByDate(String date);
